dependent models restructured
