+++ words/models/__init__.py
@@ -1,3 +1,2 @@
 from .word_models import *
 from .article_models import *
-from .article_word_models import *
