+++ words/models/article_models.py
@@ -1,6 +1,5 @@
 from django.db import models
 from .word_models import Word
-from .article_word_models import ArticleWord
 
 
 class ArticleManager(models.Manager):
@@ -91,3 +90,11 @@ class Article(models.Model):
   def __str__(self):
     return self.url
 
+
+class ArticleWord(models.Model):
+  word = models.ForeignKey(
+      Word, related_name='content', on_delete=models.CASCADE)
+  article = models.ForeignKey(
+      Article, related_name='content', on_delete=models.CASCADE)
+  count = models.IntegerField(default=0)
+  content = models.TextField(default=None)
